Remove real directory
