--- trtx/src/real/mod.rs
@@ -1,14 +1,0 @@
-//! Real TensorRT implementations
-//! No #[cfg] - this module is only compiled when mock feature is disabled
-
-pub mod builder;
-pub mod builder_config;
-pub mod cuda;
-pub mod cuda_engine;
-pub mod engine_inspector;
-pub mod host_memory;
-pub mod logger;
-pub mod network;
-pub mod onnx_parser;
-pub mod optimization_profile;
-pub mod runtime;
